requirements.txt
asgiref==3.3.4
Django==3.2.3
django-filter==2.4.0
django-jquery==3.1.0
gunicorn==20.1.0
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
whitenoise==5.2.0
